pageguard: reenable
it was disabled because it threw a warning and we wanted to make sure it
is not causing problems.
Problem was, that it tried to read some bytes of the next page in case
reads overlap, but it failed to do that because the page is not mapped.
We won't have overlapping reads, so no problem with that warning