nix: build gdata's libvmi fork
smartvmi currently depends on a fork of libvmi, where GDATA added
features (necessary for smartvmi), which are not yet upstream in the
libvmi project.
libvmi is available in nixpkgs, but GDATA's fork is not. This commit
adds a derivation to build the fork. All dependency are provided, so all
features (notable Xen and non-legacy KVM-VMI support) are enabled.